Created MailPackageConcern and refactored controller to implement concern
